Allow user to remove vote on ERP details page
This commit will allow a user to remove his/her vote from the ERP
details page. This can be used to correct a mistake or update your vote
when the page has changed.

It looks like this feature was already existing at some point, but it
was not working for at least 2 years.

- Remove a pre-existing rule that did not allow to delete a negative
Vote object when it had a comment
- Try to clean the Vote model to make it a bit easier to use.
- Try to remove as much logic from the HTML to make it easier to read /
understand. Lint the template.
- Change class to id in the HTML template has this button should be
unique
- Try to introduce the fact the a User / ERP combination should be
unique on the Vote model. We will introduce the constraint in another
commit.
- Move the tests related to the Vote to another file and improve them
(test for data quality after creation). Rename tests to make it easier
to understand.
- Add new tests on the unvote feature
- Add tracking for unvote