Translate activities. (#533)
* Translate activities.

Use modeltranslation to translate Activite name. At this point only the name has
to be translated.

* Remove weird space.

* Ignore operations in tests which manage data.

* Use slug to get the Other activity.

As name is language dependent, use slug to identify activity.

* Create the Other activity in tests.

Was previously created by a run python operation in a migration.
Those run python operations are now ignored in tests.

* Translate activities.

Use modeltranslation to translate Activite name. At this point only the name has
to be translated.