Improve search results (#551)
The goal of this commit is to improve search results by reducing the
number of results given by a specific search to the ones trully related.

- Use views.search only for the search page, creating a view dedicated
to the `search_in_municipality` page.
- Rewrite most of the code responsible for the search,
search_in_municipality and global map pages. The idea is to split the
work on the queryset and the work on the context in each view.
- The work on the queryset has different "modes" : search around me,
search in a city based on the slug, search in a street in a city, etc.
The idea behind this is to give more exact results. The way it used to
work, eveything was based on long and lat, and we searched eveything
that was in a 10Km radius. This gave you situation (example: Paris)
where you could end up with thoussands of results (mostly unrelated).
- Move tests to a dedicated file and add some more tests