Dynamic map (#587)
* Allow map to refresh data on movement

This commit will add the possibility to refesh the data on any map of
the application when the user moves the map (drag, scroll, etc).
The idea behind this is to get a more dynamic map and allow users to
look around a bit more.

We did choose to change all the maps at once to minize the change and
also to keep some views with a paginated list for SEO purposes.

This commit:
- Remove "id" to the API ERP view in GeoLoc mode, we don't really want to expose this for now. Replace with the UUID, ihis is needed to handle the opening /
closing of the popups of the what (the unique identifier used to handle
this is the UUID)
- Remove the cache on the search result view. This simplifies the
handling of the token rotation for the API.
- Allow to control if a map is dynamic or not based on a context value /
data attr parameter.
- Rewrite a good chunk of `geo.js`, I removed a few things that looked
unused (`zoomTo` for example) and tried to split some of the functions.
I also renamed what I could to make everything a bit more explicit. This
would need more work to split eveything correctly, but a lot of the
logic relies on global variables which makes this a bit harder.
- Introduce the code to handle the data refresh when the move had ended.
- Make sure the list of results is also refreshed when the map is moved.
- Make the map a bit bigger on desktop

As of today this commit comes with two major limitations:
- As the API will only sends the first 50 results, we will only display
those even if we know they are more. This could be changed with a higher
page size or an appropriate pagination / lazy-loading.
- The impact on the performance is unknown. We are not able to know how
many users will be moving the map (which will trigger the API call).
This may cause an increased number of requests on the application. I
added some debouncing on the API calls to try to limit the issue.

* Locales

---------

Co-authored-by: Marie-Laure Vernay <mlvernay@gmail.com>