Display dynamic map when no search results
This commit will display the dynamic map on the search results page when
no results are found. The idea behind is to allow the user to pan / move
/ unzoom in order to find the closest result for his/her search.

This also introduce a new button that will automatically unzoom until at
least one result is found.

In more details:
- Remove `_update_search_pager`: as far as I understood we filter
strictly on the postcode so we can never have a result that is outside
of the municipality (`commune`). I also removed a few lines of HTML in
the template that was related to this part.
- Display the map even if no results, the map is centered using lon and
lat from the context. Introduce recursion logic in order to be able to
unzoom until we find at least one results. I tested with a no result
search in Nouméa: it took some time but in the end some results were
found (in metropolitan area).
- Remove a pagination block that was duplicated