Add new api to search municipalities (#630)
We uses to use the `api-adresse.data.gouv.fr` API in order to look for
addresses and municipalities.
This API has a major issue with municipalities: it will only respond
with one postcode for a given city. Some cities have multiple postcodes
(example: Lille) which is not handled. The API will respond with the
first postcode in their datastore (59160 for Lille, 59000 is the main
postcode).
We contacted the team in charge for the API but it look like this won't
be changed / fixed even if we provided the exact root cause of the issue
(https://github.com/addok/addok/blob/master/addok/core.py#L54).

In order to make our product usable we have to switch to another API for
municipalities. We will now be using `geo.api.gouv.fr` to look for
cities. This commit implement this change:

- Add the new API with a boost on population to make sure bigger cities
are first in the list
- Only display the code "departement" instead of the postcodes to make
sure this is still readable by the user when a city has a lot of
postcodes (example : Paris)
- Make sure the backend use this code "departement" to find the correct
city in the search results