Improve performances of the sitemaps for ERPs
When looking at Sentry we can see that the sitemap for the ERPs is
having an issue: mean loading time of 1.58s and p95 of 2.16s. When
looking at this data you have to take into account that this view is
using a cache, that means that some hits are served in ~5ms (which has a
lot of impact on mean and p95).

Looking at the trace you can see that the main issue is the page making
a lot of SQL queries (around 1000 per page if my understanding is
correct - the real number can't be seen in Sentry for whatever reason).

The issue is that we need to get the `activite` of the `ERP` (or at
least the slug) to create the URL of the ERP that is used in the
sitemap.

The issue is fixed using a select_related, this could be improve if
needed by loading only the fields that are needed using `.only()`.

Add a test to show that it is scaling properly. This test only work as
is because the cache is disabled for all tests in the settings
(`setting_test.py`).

Without the fix the same test was around 23 queries (after the ERPs are
created).