Remove Popper.js
This is not used directly into our code and is already bundle with
Boostrap. It is also outdated.

It was introduced in e06f72999cd4c944121fb3afc6e158c2dbd6b10e were
Bootstrap was also added in the `index.js` file, leading me to think the
first attempt was using Popper but that Bootstrap was chosen instead
(Bootstrap was already installed).