Improve performances of multiple pages (#655)
This commit will reduce the loading time of multiple pages by reducing
the number of databases calls made to gather the data.

This optimization is composed of two parts:
- Prefetching the activity (`activite`) when gathering ERPs for a list
display. The activity is always needed at some point to display the
proper icon and/or build the link to the details page (which relies on
the activity slug)
- Removing the `has_accessibilite` field from the serializer. This field
was introduced in 007ef5326877728ce6a47c4ba4f32720a81b8caa where you can
see it was used to display a specific icon on the map (function
`createIcon` in `app.js`). As of today I found no trace of this property
being used in any HTML or Javascript. We now display the same icon for
all ERPs.

Each of these improvements reduce the number of queries mades when
handling multiple ERPs and transforming them into GeoJSON using the
`make_geojson` function.

This is the case in 4 different places:
- In the search results page. A test was added to check for scaling,
without this commit the same test would need around ~15 queries to
display the results (some base queries + 1 query for the ERPs + 1 query
per ERP for the activity + 1 query per ERP for the accesbility data).
- In the `search_in_municipality` pages. A test was also added in this
case (same results). I removed the cache because it seems the page is
now fast enough that we don't need one.
- In the ERP details pages when we show the nearest ERPS.
- In the `contrib_admin_infos` view - this case is probably quite rare.

This should improve the overall performance of the website by avoiding a
lot of queries and reducing the latency for some pages.