Add Javascript file for DSFR and fix markup
In the future we are going to start using the Js part of the French
Design System (for tabs, tags, etc).
To do so we need to add the Javascript file to the index.js file. When
doing so the DSFR will generate a JS error because our HTML markup for
the header is not valid (at least not from the DSFR point of view).

This commit fixes the markup so that we don't have this error for now.
This does not mean we have a fully fonctionnal menu / header from the
DSFR point of view as this would mean more work and some changes in the
behavior / UX of the menu.

The DSFR *needs* this specific markup to be there for the component to
work.

https://github.com/GouvernementFR/dsfr/blob/main/src/component/header/script/header/header-links.js#L12