Optimize API Key verification (#682)
* Optimize API Key verification

This commit optmize the time spent to verify an API key.

Using two classes in `DEFAULT_PERMISSION_CLASSES` was causing an
increase in latency for all API views. This is due to the fact that both
the `HasAPIKey` and `IsAllowedForAction` classes are checking for the
existence of the key using `get_from_key` (this calls a `is_valid`
method which is doing some criptographic verifications that are slow).

As the `IsAllowedForAction` class is already checking the key, we don't
really need the `HasAPIKey` class.

This commit also add a custom span so that in the future we know what is
happening in the traces (avoiding a big blank in sentry)

* Safe net if not auth header

---------

Co-authored-by: Marie-Laure Vernay <mlvernay@gmail.com>