Fix CSRF issue on homepage and remove cache (#687)
Due to the cache that was added on the homepage route, the CSRF token
for the language selection was outdated most of the time, leading to a
CSRF error.

- Fix the CSRF error by removing the cache on the route
- Add cache on the only part of the template that was a bit long
(displaying statistics, which needs some expensive DB queries)
- Rewrite the queries so that they can be kept lazy until the template
block - it seems the .count() operation is never kept lazy when passed
into the context. I passed queryset instead and moved the count
operation in the template.