Remove warnings on ERP details page
This commit will remove the "warning" ("attention") text on the ERP
details page. This was only used for screen reader and made the text
harder to underestand. The user was not sure why specific fields had a
warning and others had not.

- Rework the form to make it easier to understand. I am not sure using a
form to display only data (in a non-form interface) is a good idea.
- Make sure the logic is in the python only to avoid some logic ending
in the template. The buisiness rules are in the python code and the HTML
will display everything that is in the context.
- Clean the HTML as some parts were unused.
- Remove the flatten parameter from `get_accessibilite_data` and rewrite
tests accordingly.