Remove caching on all URLs (#688)
This commit removes all the remaining caching at the route / url level.
Most of them were not needed and introduced the same CSRF bug that was
fixed in 9e293c6e1a02171304ed210986e55612a44a1bf5 .

In order to make sure the stats page is still fast enough I rewrote a
query to make it faster. Using the ERP model to get the number of
contributors can seem counter-intuitive but it avoids a JOIN operation
between the user and erp tables by using the fact that unique users is
equivalent to unique user ids.
This change make this query around 5 times faster. Further work could be
done by moving some of the data to the GlobalStats model.