Always use bounding box on search results
The idea is to always call the bounding box API for search results.

This will "fix" the following scenario:

- A user searches "restaurant" in Paris and gets 100 results (via the
backend directly in the context)
- The user adds a filter and remove this filter and now gets 120 results
(via the API)

This was due to the fact that the 2 ways of communication are not
looking for results in the same way. The backend did a search on the
city, the API used the bounding box of the map to do the search.