Allow to have multiple widgets on the same page
This commit will allow third-parties to host multiple acceslibre widgets
on the same page. That was not possible as most of the code made the
assumption only one widget was loaded on every page.
The main challenge of this commit is to keep the retro-compatibility
with widgets that are already everywhere on internet.

- Rewrite a good chunk of widget.js
- Scope all the Javascript to make sure there is side effect on the
hosting website
- Make sure the link opens the correct modal and that everything still
works (calling matomo, closing the modal, etc).

Known issues:
- Clicking outsite of the modal won't close it (as it was already the
case)
- All the environnements are calling the same Matomo instance