Make sure widget tracking is scoped (#733)
The purpose of this commit is to make sure our tracking with Matomo for
the widget usage stay consistent when the host website also uses Matomo
as a tracking solution.

We ran into (at least) one case were the host website was using Matomo
as a tracking solution. Depending on how the tracking is implemented the
events can be sent to the wrong Matomo server due to the fact everone is
using the same `default` implementation.

This can be tested in the following way:

- Create a page with a widget
- Add the following snippet

```html
<script type="text/javascript">
  (function() {
  var _paq = window._paq = window._paq || [];
  _paq.push(['trackPageView']);
  _paq.push(['enableLinkTracking']);
  (function() {
    var u="https://demo.matomo.cloud/";
    _paq.push(['setTrackerUrl', u+'matomo.php']);
    _paq.push(['setSiteId', 1]);
    var d=document, g=d.createElement('script'), s=d.getElementsByTagName('script')[0];
    g.type='text/javascript'; g.async=true; g.src=u+'matomo.js'; s.parentNode.insertBefore(g,s);
  })();
  })();
</script>
```

- Trigger an event (by opening or closing the modal)

Before this commit, this procedure would have sent the event to
demo.matomo.cloud. With this commit the event is sent to our own Matomo
instance.

I also changed the way Matomo is implemented by extracting it from the
loop, to repeating the same code over and over.

https://developer.matomo.org/guides/tracking-javascript-guide#multiple-piwik-trackers