Do not force submit button to be enabled.
Let the `activateSubmitBtn` function responsible for checking if the button
should be enabled or not.
If provided address is not valid, `code` will remain empty and so, we should not
force the button to be enabled.

Remove spinner. Seems useless, if it need to be bring back, it should be done in
a better manner.

Move the `activateSubmitBtn` in onSubmit, so that if search data is emptied it
will not activate the button.

The 2 scenarios to test :
- an invalid address should not activate the button
- choosing an entry with the keyboard (enter key) should activate the button