Avoid creation of revision when no changes on ERP (#736)
* Avoid creation of revision when no changes on ERP

This commit will avoid the creation of a new revision on the ERP object,
when it's edited without any changes in the first step of the
contribution edit flow ("infos" step)

This is the first step to make sure that going thought the edit flow
without any change does not trigger the creation of a revision (more
work needs to be done in this direction).

- Add django-webtest to make sure it is easier to test forms /
contribution flows
- Add more fixture to ease testing
- Make sure we call the .save() method only when some changes were
actually made:
    - Use Decimal for the initial data, this will be used in the Django
internal to determine if the data has changed (`has_changed` method of
the field object). As the field use the form is a DecimalField the
cleaned data will be using a Decimal type. In order for the
comparison to check for equality properly we need the types to be
identical
    - Check for changes in the form using the `has_changed` method of
the form
- Add id in HTML to ease testing
- Add tests on the two cases we consider (edit an ERP without
modifications, edit an ERP with modifications)

* More flake8 due to upgrade

---------

Co-authored-by: Marie-Laure Vernay <mlvernay@gmail.com>