Manage widget template for non found ERPs.
We cannot raise a 404 error at this point, it would contamined the host site.
Create a distinct template for non found ERPs.
Keep the given uuid in widget tag as a unique reference to be able to close the
modal and so on.
Manage the case when data-pk in tag is not a valid UUID, handle this as a 404 /
ERP not found.
Fix ACCESLIBRE-15H