Add autocomplete for activity select (#863)
The purpose of this commit is to change the select input for the ERP
activity to a field with an autocomplete on the name and the keywords of
the activity.

- Write a custom field that will allow to re-use this on other pages
(more on this later).
- I had to move the declaration of the field to the __init__ class to
avoid issues with Python trying to access the DB before the tables are
created in the migrate script of the CI. By moving the field to the
__init__ the call to the DB is made later in the process (only when the
form is really called and not when the file / module is loaded).
- Move the logic from ActivitySelect.js to the onSubmit function of the
autocomplete library we are using).

Later on we should be re-using this component on the second page of the
contrib process in order to remove the select2 / selectwoo library from
the code.

https://autocomplete.trevoreyre.com/#/javascript-component