Add newsletter opt-in on user (#909)
Add a boolean field on the user preferences to know if the user has
opted-in to the newsletter (default to false in the model but also in
the registration form).

- Add a checkbox in the registration form to opt-in to the newsletter. I
had to rewrite the view because the change on user preference could not
be done in the form save() method (the user is not created at this point
because commit is False).
- Add checkbox to opt-in / opt-out in the user preferences. I rewrote
the form handling as it was not taking advantage of the fact that the
form is a ModelForm. Any new field that will be added in the future will
be taken into account without further modification.
- Send the data to Brevo