Adapt command to remove duplicates for town halls (#1025)
This change will lead to the number of "unhandled" cases to be falsy in
dry mode because all the ERPs will be counted (see test).

As we are no longer sure that we keep the ERP coming from service public
we need to copy the ASP Id when it exists.