Add script to create and update municipalities (#1047)
Using the JSON file in order to make the process faster than using the
API. The JSON file also contains the links between old municipalities
and new ones (which is not the case for the API).

- Write a command that will create new municipalities. Note that I had
to get the geometry from the API to be able to create the municipality.
- Update the existing municipalities (without touching geometry and
contour, this will be done in a separate script).
- Remap existing ERPs to new municipalities when possible
- Remove previous command that was using an outdated source