Fix the "having label" filter
Fix the having label manager method, this will fix the corresponding
having label filter in the search (and in the API).

The case that was not handled is when the labels array is set to an
empty list ([]). This situation was not excluded by the "isnull" clause.
Switch to using the length of the array and add tests.