Change translations and images in contribution (#1141)
One of the feedback was that the "Je ne suis par sûr" text was no
inclusive. I changed the text, but did not change the alt text of the
answer to avoid adding more complexity for screen readers. I took this
choice as "sûr" and "sûre" have the same pronunciation.

https://www.lalutineduweb.fr/ecriture-inclusive-accessibilite-numerique-jetdv2020/