Always sync user to Brevo (#1123)
* Always sync user to Brevo

When following the nominal sign up funnel, while creating an ERP, the ERP is
first created with no user, then the user is created and associated to the ERP.
In that case we were not synch'ing user to Brevo as we were only synch'ing user
when ERP is created.
Also fixing the nb_erp_edited, increment counter for the request.user (was
previously incrementing the counter for erp.user :facepalm:)

* Always sync user to Brevo

When following the nominal sign up funnel, while creating an ERP, the ERP is
first created with no user, then the user is created and associated to the ERP.
In that case we were not synch'ing user to Brevo as we were only synch'ing user
when ERP is created.
Also fixing the nb_erp_edited, increment counter for the request.user (was
previously incrementing the counter for erp.user :facepalm:)