Remove matomo trigger on widget link display
In order to reduce the number of event sent to the matomo server, we
remove the event sent when the link is displayed and move the call to
the page load event only when a link is clicked.
Also remove the stats.data matomo in order to keep only one instance.