Fix ERP notifications changed. (#1227)
* Cast to str to render lazy translations.

If not casted, it remains as a str proxy function and can not be serialized. The
serialization is needed in our case when we send a mail using brevo.

* Notify only once per erp

The get_history method will always return the whole erp history list, so we have
to notify only once per subscribed user.
Pop the user from notification so that we have only serialisable items in
notification.