Introduce new display of access data on ERP detail (#1255)
Add a new way of displaying the accessbility data on the ERP details
page to try and make it more readable.

The main change we have here is that we stop considering that the way
we store the data is the same way we should display the data to the end
user.
Now some bullet points are a built by adding the information of one, two
or three fields in the database. This is an attempt to improve the
readability of the accessbility data.