Handle None values in new ERP display
The None values where not handled correctly in most of the places, this
commit tries to fix this.
I also fixed the fact that we could display title when the section was
empty. I could not do it inside the view (we do not know what we are
going to display) and did not wanted to it in Js.

If the `:empty` selector was handled correctly in the browsers with CSS4
selectors we could have use it directly but I had to remove all the
spaces with the `spaceless` template tag to clean the HTML.

https://developer.mozilla.org/fr/docs/Web/CSS/:empty