Fix municipality view for municipality with same name
Fix a case where the page for the municipalities with the same name (for
example "Saint-Pierre" which exists in 974, 15, 31 and many more).
In this case:
- All the ERPs from the cities called Saint-Pierre where listed
- The map was centered in the middle of Africa

Fix the issue by filtering also on the departement, we are OK this may
not work if we have to cities with the same name in the same
departement.