Improve accessibility of widget for screen readers (#1414)
It looks like the content of the widget was not read by screen readers,
this commits tries to fix this.

All my tests were done using Orca and Chromium, for both a classic
widget and a "controlled" widget. For each test I tried to read the
heading (shortcut `h` for Orca) and the content of the modal (down arrow
to read line per line using Orca).

The main problem this commits fixes is that all the content of the
widget was inside a div with the `aria-hidden` set to true, even if the
content itself (nested inside this div) had `aria-hidden`set to false,
it looks like screen-readers won't consider the content in this case.
I did not found a clear answer on whether nested `aria-hidden` are valid
or not, but considered it is probably a bad idea anyway. Using my test
setup (I guess support can vary with other softwares / configuration),
nested aria attributes are just ignored and all the content is skipped
if the parent div has an `aria-hidden` attribute set to true.

This commits also fixes an image tag, and hide horizontals rules (hr)
elements for screen readers (the structure of the html should be
enough).

- https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Attributes/aria-hidden
- https://stackoverflow.com/questions/21828152/nested-aria-hidden (no
  source)