Hide empty section on ERP page for old browsers
The `has(+ ul:empty)` CSS rule is not handled properly on old browsers
(including those used in the french administration) so we are adding
this JavaScript logic to hide the title when needed.
I am also leaving the CSS rule is anyone don't want to use the
JavaScript but has a decent browser.