Change content of stats page
The purpose of this commit is to update the content of the stats page.
The data is a mixed of data from the database, from matomo and from
metabase widgets.
Starting from this commit we won't be drawing a chart on our own, this
will lead to the deletion of the chart.js dependency.