Upgrade django-registration, new button to activate account.
Upgrade to django-registration 5.1.0 which comes with some breaking changes.
We now have a new extra step: the link sent via email to activate the account,
is now redirection to a page with a button to click on.
The activation_key is now a GET parameter.
The new extra step requires a new template named activation_form.html
Clean CSS on buttons.